+++ src/components/pages/LoadMorePage/CategoryPage/index.tsx
@@ -1,0 +1,5 @@
+import LoadMoreCategoryContent from '@/components/pages/LoadMorePage/CategoryPage/LoadMoreCategoryContent';
+
+export default function LoadMoreCategoryPage() {
+  return <LoadMoreCategoryContent />;
+}
