+++ src/components/pages/LoadMorePage/PopularPage/index.tsx
@@ -1,0 +1,5 @@
+import LoadMorePageContent from '@/components/pages/LoadMorePage/PopularPage/LoadMorePageContent';
+
+export default function LoadMorePage() {
+  return <LoadMorePageContent />;
+}
