+++ src/components/pages/LoadMorePage/RandomCategoryPage/index.tsx
@@ -1,0 +1,5 @@
+import LoadMoreRandomCategoryContent from '@/components/pages/LoadMorePage/RandomCategoryPage/LoadMoreRandomCategoryContent';
+
+export default function LoadMoreRandomCategoryPage() {
+  return <LoadMoreRandomCategoryContent />;
+}
