--- src/components/pages/LoadMorePage/category/index.tsx
@@ -1,5 +1,0 @@
-import LoadMoreCategoryContent from '@/components/pages/LoadMorePage/category/LoadMoreCategoryContent';
-
-export default function LoadMoreCategoryPage() {
-  return <LoadMoreCategoryContent />;
-}
