--- src/components/pages/LoadMorePage/popular/index.tsx
@@ -1,5 +1,0 @@
-import LoadMorePageContent from '@/components/pages/LoadMorePage/popular/LoadMorePageContent';
-
-export default function LoadMorePage() {
-  return <LoadMorePageContent />;
-}
