--- src/components/pages/LoadMorePage/random-category/index.tsx
@@ -1,5 +1,0 @@
-import LoadMoreRandomCategoryContent from '@/components/pages/LoadMorePage/random-category/LoadMoreRandomCategoryContent';
-
-export default function LoadMoreRandomCategoryPage() {
-  return <LoadMoreRandomCategoryContent />;
-}
